+++ package.json
@@ -100,6 +100,7 @@
     "urlcat": "^3.1.0",
     "use-resize-observer": "^9.1.0",
     "uuid": "^9.0.0",
+    "vaul": "^1.1.1",
     "xstate": "^4.38.3",
     "yamljs": "^0.3.0",
     "zod": "^3.22.4"
